platiV2/requirements.txt
playwright
pyyaml
requests
pyinstaller